Shulkers
Singularly awful. For some reason, Mojang kept the original shulker for the purple variant but then decided to shade the bottom segment of all the new ones darker. Because I am not a fan of stupid inconsistencies, we are not doing that.

I'm not sure that the shulker_top_<color> block files are even used, but here they are anyway. They don't match the shulker entities, because I actually used the color picker to create them so they match Mojang's.